requirements.txt
boto3>=1.26.0
python-dotenv>=1.0.0
Flask>=3.0.0
opencv-python>=4.8.0
pytz>=2024.1
